# openapi.yaml
openapi: "3.1.0"
info:
  title: MTS IaaS Cloud
  description: |
    Управление виртуальными машинами, проектами, пользователями и сетями
    на платформе MTS IaaS Cloud. Этот файл — единый источник истины (source of truth)
    для бэкенда и фронтенда.
  version: "1.0.0"
  contact:
    name: MTS IaaS Team
servers:
  - url: http://localhost:8000
    description: Local dev
  - url: http://localhost:4010
    description: Prism mock server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
      example:
        detail: "Resource not found"
    ValidationError:
      type: object
      required: [detail]
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string

    # ── Auth / User ─────────────────────────────────
    UserRegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 1
          example: "strongpassword123"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          example: "user@example.com"
        password:
          type: string
          example: "strongpassword123"
    UserResponse:
      type: object
      required: [id, email, role, is_active]
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@example.com"
        role:
          type: string
          enum: [user, admin]
          example: "user"
        is_active:
          type: boolean
          example: true
    TokenResponse:
      type: object
      required: [access_token, token_type]
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          default: bearer
          example: "bearer"
...
    UserSearchVMItem:
      type: object
      properties:
        vm_id:
          type: string
          format: uuid
        vm_name:
          type: string
        status:
          type: string
        cpu:
          type: integer
        ram:
          type: integer
        ssd:
          type: integer
        ssh_link:
          type: string
          nullable: true
        project_id:
          type: string
          format: uuid
        project_name:
          type: string

    # ── SSH ──────────────────────────────────────────
...
  responses:
...
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: "Not authenticated"
...
    BadRequest:
      description: Invalid input (malformed HTTP, body parse error, etc.)
      content:
        text/plain:
          schema:
            type: string
          example: "Invalid HTTP request received."
        text/plain; charset=utf-8:
          schema:
            type: string
          example: "Invalid HTTP request received."
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
paths:
...
  /user/register:
    post:
      operationId: registerUser
      summary: Регистрация пользователя
      tags: [user]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterRequest"
      responses:
        "201":
          description: Пользователь зарегистрирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /user/login:
    post:
      operationId: loginUser
      summary: Аутентификация пользователя
      tags: [user]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /user/search_projects:
    get:
      operationId: searchUserProjects
      summary: Поиск VM в активных проектах
      tags: [user]
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Список VM
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSearchVMItem"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ═══════════════════════════════════════════════════
  #  VM
  # ═══════════════════════════════════════════════════
...